Uploading main function
Sorry for the lack of comments! The user is repeatedly prompted to input a math calculation, which the computer then computes. Features include: reading with and without spacing (can read '2+3' and '2 + 3'), recognizing parentheses and negative numbers, order of operations, being able to use 'e' and 'pi', logarithms, and trig functions.